Add files via upload
The raw ki-cad files for this project. As a note, this project currently does not function as intended, in the future, I intend to add a ADC so that the registers are able to read analog values. Furthermore, the reset pin on the D-Flips are not routed correctly, and the pull-high function in the circuit does not work as intended.